refactor: de-dupe block.factory; moved to shared
