+++ src/courseware/data/__factories__/index.js
@@ -1,4 +1,3 @@
-import './block.factory';
 import './courseBlocks.factory';
 import './courseMetadata.factory';
 import './sequenceMetadata.factory';
