+++ src/shared/data/__factories__/index.js
@@ -1,0 +1,1 @@
+import './block.factory';
